ValueNet用の終局状態予測関数を追加
